requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.2
django-stubs==1.13.2
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
gunicorn==20.1.0
idna==3.4
mypy==0.991
mypy-extensions==0.4.3
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.3
types-requests==2.28.11.8
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.14
whitenoise==6.2.0
